+++ src/controllers/Users.ts
@@ -1,4 +1,4 @@
-import { Request, Response } from "express";
+import { Request, Response, Router } from "express";
 import Users from "../models/mysql/Users";
 import bcrypt from 'bcrypt';
 import { User } from "../models/User";
@@ -7,48 +7,83 @@ import { admin } from "../models/config";
 import jwt from "jsonwebtoken";
 import { SECRET_JWT_KEY } from "../models/config";
 export const getUsers = async (req: Request, res: Response) => {
-    const { email } = req.params;
-    if(email != admin){
-        const listUsers = await Users.findAll();
-        let users: PublicUser[] = [];
-        if(listUsers){
-            users = listUsers.map(user => user.toJSON() as PublicUser);
+    let tokenAux = req.cookies.admin_token;
+    let access = req.cookies.access_token
+    if(access && tokenAux){
+        const { email } = req.params;
+        if(email != admin){
+            const listUsers = await Users.findAll();
+            let users: PublicUser[] = [];
+            if(listUsers){
+                users = listUsers.map(user => user.toJSON() as PublicUser);
+            }
+            res.json(users);
+        }else if(email == admin){
+            const listUsers = await Users.scope('withAll').findAll();
+            let users: User[] = [];
+            if(listUsers){
+                users = listUsers.map(user => user.toJSON() as User);
+            }
+            res.json(users);
+        }else{
+            res.status(501).json({message: "Bad Request for users"});
         }
-        res.json(users);
-    }else if(email == admin){
-        const listUsers = await Users.scope('withAll').findAll();
-        let users: User[] = [];
-        if(listUsers){
-            users = listUsers.map(user => user.toJSON() as User);
-        }
-        res.json(users);
     }else{
-        res.status(501).json({message: "Bad Request for users"});
+        res.send('Ruta protegida');
     }
 }
 
 export const getUser = async (req: Request, res: Response) => {
-    const { id, email } = req.params;
-    if(email != admin){
-        const UserAux = await Users.findByPk(id);    
-        if(UserAux){
-            res.json(UserAux);
-        } else {
-            res.status(404).json({message: 'Error, User not found'})
+    let tokenAux = req.cookies.access_token;
+    let user: PublicUser = new PublicUser('','','','','');
+    if(tokenAux){
+        let userAux = await getToken(tokenAux);
+        if(userAux){
+            user = userAux;
+        }
+        const { id } = req.params;
+        if(user.email === admin){
+            const UserAux = await Users.scope('withAll').findByPk(id);    
+                if(UserAux){
+                    res.json(UserAux);
+                } else {
+                    res.status(404).json({message: 'Error, User not found'})
+                }
+        }else{
+            if(user.id === id){
+                const UserAux = await Users.scope('withAll').findByPk(id);    
+                    if(UserAux){
+                        res.json(UserAux);
+                    } else {
+                        res.status(404).json({message: 'Error, User not found'})
+                    }
+            }
         }
     }else{
-        const UserAux = await Users.scope('withAll').findByPk(id);    
-    if(UserAux){
-        res.json(UserAux);
-    } else {
-        res.status(404).json({message: 'Error, User not found'})
+        res.send('Ruta protegida');
     }
+}
+
+async function getToken(tokenAux: any){
+    let user: PublicUser = new PublicUser('','','','','');
+    try{
+        const data = jwt.verify(tokenAux, SECRET_JWT_KEY);
+        if (typeof data === 'object' && data !== null) {
+            user = data as PublicUser; // Casting si estás seguro que data contiene propiedades de User
+            return user;
+        }else{
+            return null;
+        }
+    }catch(error){
+        return null;
     }
 }
 
 export const getUserByEmail = async (req: Request, res: Response) => {
-    const { email, loggedEmail } = req.params;
-    if(loggedEmail != admin){
+    let tokenAux = req.cookies.admin_token;
+    let access = req.cookies.access_token
+    if(access && tokenAux){
+        const { email } = req.params;
         const UserAux = await Users.scope('withAll').findOne({where: {email: email}});   
         if(UserAux){
             res.json(UserAux);
@@ -56,23 +91,7 @@ export const getUserByEmail = async (req: Request, res: Response) => {
             res.status(404).json({message: 'Error, User not found'})
         }
     }else{
-        const UserAux = await Users.scope('withAll').findOne({where: {email: email}});    
-        if(UserAux){
-            res.json(UserAux);
-        } else {
-            res.status(404).json({message: 'Error, User not found'})
-        }
-    }
-}
-export const temporalLogin = async (req: Request, res: Response) => {
-    const { email, password } = req.body;
-    const userAux = await loginCheck(email, password);
-    let userValidated: User = new User('','','','',false,'','');
-    if(userAux != null){
-        userValidated = userAux;
-        res.json(userValidated);
-    }else{
-        res.status(404).json({message: 'Error al iniciar sesion'});
+        res.send('Ruta protegida');
     }
 }
 
@@ -86,12 +105,26 @@ export const login = async (req: Request, res: Response) => {
             expiresIn: "1h"
         });
 
+        const refresh_token = jwt.sign({ id: userValidated.id, email: userValidated.email, priceList: userValidated.priceList, username: userValidated.username, client: userValidated.client }, SECRET_JWT_KEY, {
+            expiresIn: "1d"
+        });
+
         res.cookie('access_token', access_token, {
             path: '/',
             httpOnly: true,
             secure: true,///process.env.NODE_ENV == 'production',
             sameSite: 'none',
+            domain: '.somostresbe.com', // Comparte la cookie entre www.somostresbe.com y api.somostresbe.com
             maxAge: 1000 * 60 * 60
+        });
+
+        res.cookie('refresh_token', refresh_token, {
+            path: '/',
+            httpOnly: true,
+            secure: true,///process.env.NODE_ENV == 'production',
+            sameSite: 'none',
+            domain: '.somostresbe.com', // Comparte la cookie entre www.somostresbe.com y api.somostresbe.com
+            maxAge: 1000 * 60 * 60 * 24
         });
 
         if(userValidated.email == admin){
@@ -119,13 +152,24 @@ export const logout = async (req: Request, res: Response) => {
     
     if(token){
         const admin = req.cookies.admin_token;
+        res.cookie('refresh_token', '', {
+            path: '/',
+            httpOnly: true,
+            secure: true,///process.env.NODE_ENV == 'production',
+            sameSite: 'none',
+            domain: '.somostresbe.com', // Comparte la cookie entre www.somostresbe.com y api.somostresbe.com
+            maxAge: 0
+        });
+
         res.cookie('access_token', '', {
             path: '/',
             httpOnly: true,
             secure: true,///process.env.NODE_ENV == 'production',
             sameSite: 'none',
+            domain: '.somostresbe.com', // Comparte la cookie entre www.somostresbe.com y api.somostresbe.com
             maxAge: 0
         });
+
         if(admin){
             res.cookie('admin_token', '', {
                 path: '/',
@@ -156,27 +200,42 @@ async function loginCheck(email: string, password: string){
 }
 
 export const getUsersBySeller = async (req: Request, res: Response) => {
-    const { seller } = req.params;
+    let tokenAux = req.cookies.admin_token;
+    let access = req.cookies.access_token;
+    if(access && tokenAux){
+        const { seller } = req.params;
     const UserAux = await Users.findAll({where: {seller: seller}});    
     if(UserAux){
         res.json(UserAux);
     } else {
         res.status(404).json({message: 'Error, User not found'})
     }
+    }else{
+        res.send('Ruta protegida');
+    }
 }
 
 export const getUserByName = async (req: Request, res: Response) => {
-    const { username } = req.params;
-    const UserAux = await Users.findOne({where: {username: username}});    
+    let access = req.cookies['access_token'];
+    let tokenAux = req.cookies['admin_token'];
+    if(access && tokenAux){
+        const { username } = req.params;
+    const UserAux = await Users.findOne({where: {username: username}});  
     if(UserAux){
         res.json(UserAux);
     } else {
         res.status(404).json({message: 'Error, User not found'})
     }
+    }else{
+        res.send('Ruta protegida');
+    }
 }
 
 export const deleteUser = async (req: Request, res: Response) => {
-    const { id } = req.params;
+    let tokenAux = req.cookies.admin_token;
+    let access = req.cookies.access_token;
+    if(access && tokenAux){
+        const { id } = req.params;
     const UserAux = await Users.findByPk(`${id}`);
     if(UserAux){
         await UserAux.destroy();
@@ -184,8 +243,14 @@ export const deleteUser = async (req: Request, res: Response) => {
     } else{
         res.status(404).json({message: 'Error, User not found'})
     }
+    }else{
+        res.send('Ruta protegida');
+    }
 }
 export const postUser = async(req: Request, res: Response) => {
+    let tokenAux = req.cookies.admin_token;
+    let access = req.cookies.access_token;
+    if(access && tokenAux){
     let {id, email, username, priceList, client, seller, password} = req.body;
     let hashedPassword = await bcrypt.hash(password, 10);
     const user = {
@@ -200,12 +265,16 @@ export const postUser = async(req: Request, res: Response) => {
     await Users.create(user);
     res.json({
         message: `User successfully created with hashed password: ${user}`,
-    })
+    });
+    }else{
+        res.send('Ruta protegida');
+    }
 }
 export const updateUser = async(req: Request, res: Response) => {
+    /*let tokenAux = req.cookies.admin_token;
+    let access = req.cookies.access_token;*/
     const body = req.body;
     const { id } = req.params;
-    body.password = await bcrypt.hash(body.password, 10);
     const UserAux = await Users.findByPk(id);
     if(UserAux){
         UserAux.update(body);
@@ -215,7 +284,35 @@ export const updateUser = async(req: Request, res: Response) => {
     } else {
         res.status(404).json({message: 'Error, User not found'})
     }
+
+}
+
+export const updatePassword = async(req: Request, res: Response) => {
+    let tokenAux = req.cookies.admin_token;
+    let access = req.cookies.access_token;
+    if(access && tokenAux){
+    const body = req.body;
+    const { id } = req.params;
+    const UserAux = await Users.findByPk(id);
+    if(UserAux){
+        body.password = await bcrypt.hash(body.password, 10);
+        UserAux.update(body);
+        res.json({
+            message: 'User updated',
+        })
+    } else {
+        res.status(404).json({message: 'Error, User not found'})
+    }
+    }else{
+        res.send('Ruta protegida');
+    }
 }
 export const deleteUsers = async (req: Request, res: Response) => {
-    await Users.destroy({truncate: true});
+    let tokenAux = req.cookies.admin_token;
+    let access = req.cookies.access_token;
+    if(access && tokenAux){
+        await Users.destroy({truncate: true});
+    }else{
+        res.send('Ruta protegida');
+    }
 }
